Move byte-to-unsigned functionality into child crate
