+++ raw/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "tinystr-raw"
+description = """
+Raw string-to-integer conversions for tinystr.
+"""
+version = "0.1.0"
+authors = ["Zibi Braniecki <zibi@braniecki.net>", "Shane Carr <shane@sffc.xyz>"]
+edition = "2018"
+license = "Apache-2.0/MIT"
+repository = "https://github.com/zbraniecki/tinystr"
+readme = "README.md"
+keywords = ["string", "str", "small", "tiny", "no_std"]
+categories = ["data-structures"]
